### 2018-06-10 #771 @whedon
[PRE REVIEW]: ggeffects - Tidy Data Frames of Marginal Effects from Regression Models
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/strengejacke/ggeffects" target ="_blank">https://github.com/strengejacke/ggeffects</a>
**Version:** 0.3.4
