### 2018-06-10 #771 @leeper
@cdcrabtree Would you be willing to review this for JOSS?
