### 2018-06-10 #771 @leeper
@mdogucu Would you be willing to review this for JOSS?
