### 2018-06-10 #771 @cdcrabtree
@leeper I would, but I am traveling and probably can't begin on this until the weekend. Does that work for JOSS?
