### 2018-06-10 068fbd3
update paper figures and captions
