### 2018-06-10 61b1da5
typo in figure legend
