### 2018-06-10 #771 @strengejacke
(I resized figures, so page breaks hopefully look better)
