### 2018-06-11 #772 @leeper
@cdcrabtree Here's the review issue. Please work through the review checklist, open issues [on the source repo](https://github.com/strengejacke/ggeffects) for any issues not addressed by the checklist or that you'd like to see changed, then send me a message here when you'd like me to have a look.
