### 2018-06-13 4e3be0b
add CONTRIBUTING.md for JOSS
