+++ README.md
@@ -89,7 +89,7 @@ There are some more features, which are explained in more detail in the package-
 
 ## Contributing to the package
 
-Please follow [this guide](CONTRIBUNTING.md) if you like to contribute to this package.
+Please follow [this guide](CONTRIBUTING.md) if you like to contribute to this package.
 
 ## Documentation
 
