### 2018-06-14 #772 @strengejacke
Is it possible to get a template for compiling the PDF? I would like to add a page break to layout the figures in the paper, and don't want to spam this thread with try&error-submission for my `paper.md` to check when the PDF looks best... So I want to check this locally. Or is there a kind of "typesetting" before publishing once a paper is accepted?
