+++ NAMESPACE
@@ -5,7 +5,6 @@ S3method(plot,ggeffects)
 S3method(plot,ggeffectslist)
 S3method(summary,ggeffects)
 export("%>%")
-export(ame)
 export(eff)
 export(emm)
 export(get_complete_df)
@@ -21,7 +20,6 @@ export(ggeffect)
 export(gginteraction)
 export(ggpoly)
 export(ggpredict)
-export(mem)
 importFrom(dplyr,arrange)
 importFrom(dplyr,bind_cols)
 importFrom(dplyr,bind_rows)

[…]
