### 2018-06-15 6c7c26c
improve plot() method
