+++ R/getter.R
@@ -49,6 +49,8 @@
 #'
 #' @export
 get_title <- function(x, case = NULL) {
+  if (sjmisc::is_empty(x)) return(NULL)
+
   if (!inherits(x, "ggeffects"))
     stop("`x` must be of class `ggeffects`.", call. = F)
 
@@ -59,6 +61,8 @@ get_title <- function(x, case = NULL) {
 #' @rdname get_title
 #' @export
 get_x_title <- function(x, case = NULL) {
+  if (sjmisc::is_empty(x)) return(NULL)
+
   if (!inherits(x, "ggeffects"))
     stop("`x` must be of class `ggeffects`.", call. = F)
 
@@ -69,6 +73,8 @@ get_x_title <- function(x, case = NULL) {
 #' @rdname get_title
 #' @export
 get_y_title <- function(x, case = NULL) {
+  if (sjmisc::is_empty(x)) return(NULL)
+
   if (!inherits(x, "ggeffects"))
     stop("`x` must be of class `ggeffects`.", call. = F)
 
@@ -79,6 +85,8 @@ get_y_title <- function(x, case = NULL) {
 #' @rdname get_title
 #' @export
 get_legend_title <- function(x, case = NULL) {
+  if (sjmisc::is_empty(x)) return(NULL)
+
   if (!inherits(x, "ggeffects"))
     stop("`x` must be of class `ggeffects`.", call. = F)
 
@@ -89,6 +97,8 @@ get_legend_title <- function(x, case = NULL) {
 #' @rdname get_title
 #' @export
 get_legend_labels <- function(x, case = NULL) {
+  if (sjmisc::is_empty(x)) return(NULL)
+
   if (!inherits(x, "ggeffects"))
     stop("`x` must be of class `ggeffects`.", call. = F)
 
@@ -99,6 +109,8 @@ get_legend_labels <- function(x, case = NULL) {
 #' @rdname get_title
 #' @export
 get_x_labels <- function(x, case = NULL) {
+  if (sjmisc::is_empty(x)) return(NULL)
+
   if (!inherits(x, "ggeffects"))
     stop("`x` must be of class `ggeffects`.", call. = F)
 
