+++ NAMESPACE
@@ -2,7 +2,6 @@
 
 S3method(plot,ggalleffects)
 S3method(plot,ggeffects)
-S3method(plot,ggeffectslist)
 S3method(summary,ggeffects)
 export("%>%")
 export(eff)
