### 2018-06-17 2e0531b
more options for `plot()`-method, add vignettes and tests
