+++ R/plot.R
@@ -4,47 +4,51 @@
 #' @description A generic plot-method for \code{ggeffects}-objects.
 #'
 #' @param x An object of class \code{ggeffects}, as returned by the functions
-#'          from this package.
+#'   from this package.
 #' @param ci Logical, if \code{TRUE}, confidence bands (for continuous variables
-#'          at x-axis) resp. error bars (for factors at x-axis) are plotted.
-#'          For \code{ggeffects}-objects from \code{ggpredict()} with argument
-#'          \code{full.data = TRUE}, \code{ci} is automatically set to \code{FALSE}.
+#'   at x-axis) resp. error bars (for factors at x-axis) are plotted.
+#'   For \code{ggeffects}-objects from \code{ggpredict()} with argument
+#'   \code{full.data = TRUE}, \code{ci} is automatically set to \code{FALSE}.
 #' @param facets Logical, defaults to \code{TRUE}, if \code{x} has a column named
-#'          \code{facet}, and defaults to \code{FALSE}, if \code{x} has no such
-#'          column. Set \code{facets = TRUE} to wrap the plot into facets even
-#'          for grouping variables (see 'Examples').
+#'   \code{facet}, and defaults to \code{FALSE}, if \code{x} has no such
+#'   column. Set \code{facets = TRUE} to wrap the plot into facets even
+#'   for grouping variables (see 'Examples').
 #' @param rawdata Logical, if \code{TRUE}, a layer with raw data from response by
-#'          predictor on the x-axis, plotted as point-geoms, is added to the plot.
+#'   predictor on the x-axis, plotted as point-geoms, is added to the plot.
 #' @param colors Character vector with color values in hex-format, valid
-#'          color value names (see \code{demo("colors")} or a name of a
-#'          \href{http://colorbrewer2.org}{color brewer} palette.
-#'          Following options are valid for \code{colors}:
-#'          \itemize{
-#'            \item If not specified, the color brewer palette "Set1" will be used.
-#'            \item If \code{"gs"}, a greyscale will be used.
-#'            \item If \code{"bw"}, the plot is black/white and uses different line types to distinguish groups.
-#'            \item If \code{colors} is any valid color brewer palette name, the related palette will be used. Use \code{\link[RColorBrewer]{display.brewer.all}} to view all available palette names.
-#'            \item Else specify own color values or names as vector (e.g. \code{colors = c("#f00000", "#00ff00")}).
-#'          }
+#'   color value names (see \code{demo("colors")} or a name of a
+#'   \href{http://colorbrewer2.org}{color brewer} palette.
+#'   Following options are valid for \code{colors}:
+#'   \itemize{
+#'     \item If not specified, the color brewer palette "Set1" will be used.
+#'     \item If \code{"gs"}, a greyscale will be used.
+#'     \item If \code{"bw"}, the plot is black/white and uses different line types to distinguish groups.
+#'     \item If \code{colors} is any valid color brewer palette name, the related palette will be used. Use \code{\link[RColorBrewer]{display.brewer.all}} to view all available palette names.
+#'     \item Else specify own color values or names as vector (e.g. \code{colors = c("#f00000", "#00ff00")}).
+#'   }
 #' @param alpha Alpha value for the confidence bands.
 #' @param dodge Value for offsetting or shifting error bars, to avoid overlapping.
[…]
