### 2018-06-17 c87c26d
add tests for various arguments
