### 2018-06-17 bac72d7
add more tests
