+++ R/ggpredict.R
@@ -17,31 +17,20 @@
 #'   three terms, where the second and third term indicate the groups, i.e.
 #'   predictions of first term are grouped by the levels of the second (and third)
 #'   term. If \code{terms} is missing or \code{NULL}, marginal effects for each
-#'   model term are calculated. Indicating levels in square brackets allows for
-#'   selecting only specific groups. Term name and levels in brackets must be
-#'   separated by a whitespace character, e.g.
-#'   \code{terms = c("age", "education [1,3]")}. Numeric ranges, separated
-#'   with colon, are also allowed: \code{terms = c("education", "age [30:60]")}.
-#'   Furthermore, it is possible to specify a function name. Values for
-#'   predictions will then be transformed, e.g. \code{terms = "income [exp]"}.
-#'   This is useful when model predictors were transformed for fitting the
-#'   model and should be back-transformed to the original scale for predictions.
-#'   See 'Examples' and package vignettes. All remaining covariates that are
-#'   not specified in \code{terms} are held constant (if \code{full.data = FALSE},
-#'   the default) or are set to the values from the observations (i.e. are kept
-#'   as they happen to be; see 'Details').
+#'   model term are calculated. It is also possible to define specific values for
+#'   terms, at which marginal effects should be calculated (see 'Details').
+#'   All remaining covariates that are not specified in \code{terms} are held
+#'   constant (if \code{full.data = FALSE}, the default) or are set to the
+#'   values from the observations (i.e. are kept as they happen to be;
+#'   see 'Details'). See also argument \code{conditon} and \code{typical}.
 #' @param ci.lvl Numeric, the level of the confidence intervals. For \code{ggpredict()},
 #'   use \code{ci.lvl = NA}, if confidence intervals should not be calculated
 #'   (for instance, due to computation time).
 #' @param type Character, only applies for mixed effects models. Indicates
 #'   whether predicted values should be conditioned on random effects
 #'   (\code{type = "re"}) or fixed effects only (\code{type = "fe"},
-#'   the default). If \code{type = "re"}, the uncertainty in the variance
-#'   parameters is currently not taken into account. The specific levels of
-#'   the random effects are not provided in the \code{newdata}-argument. Hence,
-#'   if \code{type = "re"}, \code{predict()} is called with \code{re.form = NULL},
-#'   i.e. predictions are conditioned on random effects only using the first level
-#'   of random effects, but the uncertainty in the variance parameters is ignored.
+#'   the default). If \code{type = "re"}, prediction intervals also consider
+#'   the uncertainty in the variance parameters.
 #' @param full.data Logical, if \code{TRUE}, the returned data frame contains
 #'   predictions for all observations. This data frame also has columns
 #'   for residuals and observed values, and can also be used to plot a
@@ -79,6 +68,7 @@
 #' @param ... Further arguments passed down to \code{predict()}.
 #'
 #' @details
+#'   \strong{Supported Models} \cr \cr
 #'   Currently supported model-objects are: \code{lm, glm, glm.nb, lme, lmer,
 #'   glmer, glmer.nb, nlmer, glmmTMB, gam, vgam, gamm, gamm4, multinom,
 #'   betareg, gls, gee, plm, lrm, polr, clm, hurdle, zeroinfl, svyglm,
@@ -91,14 +81,32 @@
 #'   \code{ggeffect()} computes a kind of "average" value, which represents
 #'   the proportions of each factor's category.
 #'   \cr \cr
+#'   \strong{Marginal Effects at Specific Values} \cr \cr
[…]
