+++ vignettes/plotmethod.Rmd
@@ -16,6 +16,8 @@ options(width = 800, tibble.width = Inf)
 
 # plot()-method
 
+This vignettes demonstrates the `plot()`-method of the **ggeffects**-package. It is recommended to read the [general introduction](marginaleffects.html) first, if you haven't done this yet.
+
 If you don't want to write your own ggplot-code, **ggeffects** has a `plot()`-method with some convenient defaults, which allows quickly creating ggplot-objects. `plot()` has only a few arguments to keep this function small and simple. For instance, `ci` allows you to show or hide confidence bands (or error bars, for discrete variables), `facets` allows you to create facets even for just one grouping variable, or `colors` allows you to quickly choose from some color-palettes, including black & white colored plots. Use `rawdata` to add the raw data points to the plot.
 
 ```{r}
