+++ vignettes/marginaleffects.Rmd
@@ -90,56 +90,6 @@ mydf <- ggaverage(fit, terms = c("c12hour", "c172code"))
 ggplot(mydf, aes(x, predicted, colour = group)) + geom_line()
 ```
 
-# Marginal effects at specific values or levels
-
-The `terms`-argument not only defines the model terms of interest, but each model term can be limited to certain values. This allows to compute and plot marginal effects for (grouping) terms at specific values only, or to define values for the main effect of interest.
-
-There are several options to define these values, which always should be placed in square brackets directly after the term name and can vary for each model term.
-
-1. Concrete values are separated by a comma: `terms = "c172code [1,3]"`. For factors, you could also use factor levels, e.g. `terms = "Species [setosa,versicolor]"`.
-2. Ranges are specified with a colon: `terms = c("c12hour [30:80]", "c172code [1,3]")`. This would plot all values from 30 to 80 for the variable _c21hour_.
-3. A function name. The function is then applied to all unique values of the indicated variable, e.g. `terms = "hp [exp]"`.
-4. A `range`, for numeric variables. In this case, `ggpredict()` tries to calculate a pretty numeric range based on how large the range of the related variable is. Example: `terms = "age [range]"`.
-
-```{r}
-mydf <- ggpredict(fit, terms = c("c12hour [30:80]", "c172code [1,3]"))
-mydf
-ggplot(mydf, aes(x, predicted, colour = group)) + geom_line()
-```
-
-Defining value ranges is especially useful when variables are, for instance, log-transformed. `ggpredict()` then typically only uses the range of the log-transformed variable, which is in most cases not what we want. In such situation, specify the range in the `terms`-argument.
-
-```{r}
-data(mtcars)
-mpg_model <- lm(mpg ~ log(hp), data = mtcars)
-
-# x-values and predictions based on the log(hp)-values
-ggpredict(mpg_model, "hp")
-
-# x-values and predictions based on hp-values from 50 to 150
-ggpredict(mpg_model, "hp [50:150]")
-```
-
-The brackets in the `terms`-argument also accept the name of a valid function, to (back-)transform predicted valued. In this example, an alternative would be to specify that values should be exponentiated, which is indicated by `[exp]` in the `terms`-argument:
-
-```{r}
-# x-values and predictions based on exponentiated hp-values
-ggpredict(mpg_model, "hp [exp]")
-```
-
-# Marginal effects conditioned on specific values of the covariates
-
-By default, the `typical`-argument determines which function will be applied to the covariates to hold these terms constant. Use the `condition`-argument to define specific values at which a covariate should be held constant. `condition` requires a named vector, with the name indicating the covariate.
-
-```{r}
-mpg_model <- lm(mpg ~ log(hp) + disp, data = mtcars)
-
-# "disp" is hold constant at its mean
-ggpredict(mpg_model, "hp [exp]")
-
-# "disp" is hold constant at value 200
-ggpredict(mpg_model, "hp [exp]", condition = c(disp = 200))
-```
 
 # Two- and Three-Way-Interactions
 
