### 2018-06-19 3f4e4e5
improve automatic axis-breaks for plot(log.y = T)
