+++ NAMESPACE
@@ -90,6 +90,7 @@ importFrom(sjmisc,to_factor)
 importFrom(sjmisc,to_long)
 importFrom(sjmisc,trim)
 importFrom(sjmisc,var_rename)
+importFrom(sjmisc,zap_inf)
 importFrom(sjstats,hdi)
 importFrom(sjstats,link_inverse)
 importFrom(sjstats,model_family)
