### 2018-06-20 dedc29b
add tests, typo in paper
