+++ tests/testthat/test-glmer.R
@@ -39,3 +39,14 @@ test_that("ggeffect, glmer", {
   ggeffect(fit, c("c12hour", "c161sex"))
   ggeffect(fit, c("c12hour", "c161sex", "c172code"))
 })
+
+library(glmmTMB)
+data(Owls)
+m <- glmer.nb(SiblingNegotiation ~ SexParent + ArrivalTime + (1 | Nest), data = Owls)
+
+test_that("ggpredict, glmer.nb", {
+  ggpredict(m, "SexParent")
+  ggpredict(m, "SexParent", type = "re")
+  ggpredict(m, c("SexParent", "ArrivalTime"))
+  ggpredict(m, c("SexParent", "ArrivalTime"), type = "re")
+})
