### 2018-06-20 706eb6a
close https://github.com/strengejacke/sjPlot/issues/377 (be more explicit about problems with logical vectors)
