### 2018-06-21 862a84f
improve automatic prettifying
