### 2018-06-22 524ba57
add support for models from pkg "robust"
