### 2018-06-22 a617533
support lmrob, glmrob and brglm models
