### 2018-06-23 70b3a98
add reference to figure in paper.md
