### 2018-06-29 bcdc528
refactoring, export "pretty_range()" function
