+++ NAMESPACE
@@ -18,6 +18,7 @@ export(ggeffect)
 export(gginteraction)
 export(ggpoly)
 export(ggpredict)
+export(pretty_range)
 importFrom(dplyr,arrange)
 importFrom(dplyr,bind_cols)
 importFrom(dplyr,bind_rows)
