### 2018-06-29 438b5af
add tests for label-getter
