### 2018-06-29 12db705
make docs more detailled, add more "terms"-options
