+++ R/ggpoly.R
@@ -20,10 +20,13 @@
 #'           \item{\code{group}}{the grouping level from the second term in \code{terms}, used as grouping-aesthetics in plots.}
 #'         }
 #'
-#' @note \code{ggpoly()} is just an alternative call to \code{ggpredict()} for
-#'       polynomial model terms. It may work for certain models that are not
-#'       yet supported by \code{ggpredict()}. Otherwise, there should be no
-#'       difference in the results from \code{ggpoly()} and \code{ggpredict()}.
+#' @details \code{ggpoly()} is just an alternative call to \code{ggpredict()} for
+#'   polynomial model terms. It internally calls \code{effect()} from \pkg{effects}
+#'   (like \code{ggeffect()} does), hence it may work for certain models that are not
+#'   yet supported by \code{ggpredict()}. Otherwise, there should be no difference
+#'   in the results from \code{ggpoly()} and \code{ggpredict()}, or at least only
+#'   minor differences (as \code{effect()} holds categorical covariates constant at
+#'   other values, see 'Details' in \link{ggeffect}).
 #'
 #' @examples
 #' data(efc)
