### 2018-06-29 e8c94e2
update vignette
