### 2018-06-29 c36ab15
add tests for new terms-option
