### 2018-06-29 164a607
release 0.3.4-1 for JOSS
