### 2018-06-29 #772 @strengejacke
Thanks, Charles, for reviewing my submission! I have created a new release that includes all reviewer suggestions and created a doi for this version, which can be used for the submitted paper:

https://doi.org/10.5281/zenodo.1301148
